Add LinkExtrasModule
