+++ paymentsheet/src/main/java/com/stripe/android/paymentsheet/injection/AddressElementViewModelFactoryComponent.kt
@@ -3,6 +3,7 @@ package com.stripe.android.paymentsheet.injection
 import android.content.Context
 import com.stripe.android.core.injection.CoreCommonModule
 import com.stripe.android.core.injection.CoroutineContextModule
+import com.stripe.android.link.injection.LinkExtrasModule
 import com.stripe.android.payments.core.injection.StripeRepositoryModule
 import com.stripe.android.paymentsheet.addresselement.AddressElementActivityContract
 import com.stripe.android.paymentsheet.addresselement.AddressElementViewModel
@@ -21,7 +22,8 @@ import javax.inject.Singleton
         CoreCommonModule::class,
         AddressElementViewModelModule::class,
         FormControllerModule::class,
-        ResourceRepositoryModule::class
+        ResourceRepositoryModule::class,
+        LinkExtrasModule::class
     ]
 )
 internal interface AddressElementViewModelFactoryComponent {
